Huge Changes
Decided the calendar functionality is pointless, made the choice to get
rid of it. It did help me learn a lot, but it's time that we did
something else.

As a benefit of removing the calendar functions, I dove in and started
working on the base program. Rewamped the way the tray icon is handeled,
put new menues in a few places, added an options screen and started
working on an overlay. It's almost done for the stopwatch.

I also added a class to make global hotkeys work, so you can now
controll the stopwatch with some keys (will let them be changed later).
I'll also mod that class a bit, so that (hopefully), I'll be able to not
only controll the coundown, but also set the time on it.