Overlay Update 2
Ripped out old overlay code, and re-wrote it form the ground up. It's
now not only cooler and more useful, but fully compatible with the g510
(at least that one g510 lib for C#).

I also changed how the hot keys work, so that instead of having a hotkey
for every function, it now only uses 4, one to cycle the "tabs" on the
overlay, and the other 3 are tab specific.

Also fixed a buch of inconsitancies with my camel case and labeling
style (haven't even touched the countdown tab class yet, so that's still
shitty).